refactor: move adding slide logic to store
